Don’t send extra fields round
