+++ packages/inquiry/routers/submit.js
@@ -1,4 +1,5 @@
 const { asyncRoute } = require('@base-cms/utils');
+const { getAsArray } = require('@base-cms/object-path');
 const { content: contentLoader } = require('@base-cms/web-common/page-loaders');
 const send = require('@endeavorb2b/base-website-common/utils/send-mail');
 const buildMarkoGlobal = require('@endeavorb2b/base-website-common/utils/build-marko-global');
@@ -13,18 +14,27 @@ module.exports = ({ queryFragment, notification, confirmation }) => asyncRoute(a
     directSend,
   } = site.getAsObject('inquiry');
   const $global = buildMarkoGlobal(res);
-  const { apollo } = req;
+  const { apollo, body: payload } = req;
   const content = await contentLoader(apollo, { id: req.params.id, queryFragment });
+  const contacts = getAsArray(content, 'inquiryContacts')
+    .map(({ name, email }) => ({ name, email }))
+    .filter(({ email }) => email);
+  const addresses = {
+    to: directSend && contacts.length ? contacts : to,
+    cc: directSend && contacts.length ? to : undefined,
+    bcc,
+    from,
+  };
+  // Notify the contacts of the submission
   await send(notificationBuilder({
     template: notification,
     $global,
     content,
-    payload: req.body,
-    to,
-    from,
-    bcc,
-    directSend,
+    payload,
+    addresses,
   }));
+
+  // Notify the user their submission was received
   if (req.body.confirmationEmail) {
     await send(confirmationBuilder({
       template: confirmation,
